# backend/src/main/java/com/daypoo/api/dto/LoginRequest.java
package com.daypoo.api.dto;

import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
    @NotBlank(message = "아이디는 필수입니다.") String username,
    @NotBlank(message = "비밀번호는 필수입니다.") String password) {}