# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: DayPoo API
  description: 💩 DayPoo - 실시간 화장실 찾기 및 건강 리포트 AI 서비스 API
  version: v1.0.0
servers:
  - url: http://localhost:8080
    description: Local Server
paths:
  /api/v1/auth/signup:
    post:
      tags:
        - Authentication
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
      responses:
        "200":
          description: OK
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SignUpRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        nickname:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...